wip: going back to the activities with fragments
