+++ jukebox-app/src/main/AndroidManifest.xml
@@ -1,9 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    package="se.qxx.android.jukebox"
-    android:versionCode="1"
-    android:versionName="1.1">
+          xmlns:tools="http://schemas.android.com/tools"
+          package="se.qxx.android.jukebox">
 
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
@@ -37,11 +35,14 @@
             android:name=".activities.ListActivity"
             android:theme="@style/Theme.AppCompat" />
 
+        <activity
+                android:name=".activities.MovieDetailActivity"
+                android:theme="@style/Theme.AppCompat" />
+
         <meta-data
             android:name="com.google.android.gms.cast.framework.OPTIONS_PROVIDER_CLASS_NAME"
             android:value="se.qxx.android.jukebox.cast.CastOptionsProvider" />
 
-        <activity android:name=".activities.MovieDetailActivity"></activity>
     </application>
 
 </manifest>
