+++ jukebox-app/src/main/java/se/qxx/android/jukebox/activities/fragments/PagerFragment.java
@@ -15,9 +15,7 @@ import se.qxx.android.jukebox.settings.JukeboxSettings;
 import se.qxx.android.tools.Logger;
 
 public class PagerFragment extends Fragment {
-    ViewPager pager;
     private JukeboxSettings settings;
-    private JukeboxFragmentAdapter adapter;
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
@@ -33,14 +31,7 @@ public class PagerFragment extends Fragment {
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
         View v = inflater.inflate(R.layout.jukebox_main_wrapper, container, false);
-        initializeView(v);
         return v;
-    }
-
-    private void initializeView(View v) {
-        pager = getActivity().findViewById(R.id.rootJukeboxViewPager);
-        adapter = new JukeboxFragmentAdapter(this.getActivity().getSupportFragmentManager(), this.getActivity());
-        pager.setAdapter(adapter);
     }
 
     public static PagerFragment newInstance() {
