+++ jukebox-app/src/main/res/layout/jukebox_list_wrapper.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="fill_parent"
+    android:layout_height="fill_parent"
+    android:gravity="center|clip_vertical|fill_horizontal"
+    android:orientation="vertical" 
+    android:id="@+id/rootJukeboxListWrapper">
+
+	<fragment
+		android:id="@+id/castMiniController"
+		android:layout_width="fill_parent"
+		android:layout_height="wrap_content"
+		android:layout_alignParentBottom="true"
+		android:visibility="gone"
+		class="com.google.android.gms.cast.framework.media.widget.MiniControllerFragment" />
+
+</RelativeLayout>
